Update networking code
Consulting the documentation again, it appears that GIOCondition '''is'''
indeed a flag, so the last modification was not the correct solution after
all - it did not work at all in Windows.

Revert to reading all data at once and fixed the warning. First tests
indicate that it works pretty well on Linux, while on Windows I noticed
some strange problems with the watch callback being disconnected without
apparent reason. Refs #21.